REFACTOR: 1.getDocuments의 as 파라미터 타입 수정 2. DATA/Helper 루트 추가 및 Utils,Helper 경로의 파일 구성 변경
FEAT:
1. getDocument 추가
2. getDocument에 필요한 Query+map 메서드 추가
