+++ GatheRunner/Data/Helpers/CollectionReference+AddQueries.swift
